learning/building_with_claude/requirements.txt
anthropic === 0.84.0
python-dotenv === 0.9.9